SCRUM-33 added event map & live updating for creation

+++ src/components/EventList.tsx
@@ -6,6 +6,8 @@ import { faPersonRunning, faSoccerBall, faUser, faTrash } from "@fortawesome/fre
 import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet';
 import { LatLng } from 'leaflet';
 import { useUserContext } from "../userContext";
+import MainMap from "./EventsMap";
+import { useWebSocket } from '../WebsocketContext.tsx';
 
 type Event = {
     location: {
@@ -73,8 +75,8 @@ const MapComponent: React.FC<MapComponentProps> = React.memo(({ title, location 
     );
 });
 
-
 export default function EventList() {
+    const { socket } = useWebSocket();
     const [events, setEvents] = useState<Event[]>([]);
     const [loading, setLoading] = useState(true);
     const { user, isTokenExpired, resetJWT } = useUserContext();
@@ -85,6 +87,21 @@ export default function EventList() {
     useEffect(() => {
         localStorage.setItem("followingEventTable", JSON.stringify(following));
     }, [following]);
+
+    useEffect(() => {
+        if (socket) {
+            socket.on('new-event', () => {
+                console.log("received new event from server.");
+                getEvents();
+            });    
+            socket.on('error', (obj) => {
+                console.log("Error with event socket: " + obj.message);
+            });   
+              return () => {
+                socket.off('new-event');
+            };
+        }
+      }, [socket]);
 
     const getEvents = async () => {
         try {
@@ -104,7 +121,7 @@ export default function EventList() {
 
     useEffect(() => {
         getEvents();
-    }, [events]);
+    }, []);
 
     const followEvent = async (event: Event, e: React.MouseEvent<HTMLButtonElement>) => {
         if (isTokenExpired()) {
@@ -140,16 +157,21 @@ export default function EventList() {
         }
     };
 
-    const deleteEvent = async (event: Event) => {
+    const deleteEvent = async (eventId: string) => {
+        if (isTokenExpired()) {
+            resetJWT()
+            window.alert('Seja je potekla, potrebna je ponovna prijava.')
+            return
+        }
         try {
-            const response = await fetch(`http://localhost:3000/events/${event._id}`, {
+            const response = await fetch(`http://localhost:3000/events/${eventId}`, {
                 method: 'DELETE',
                 headers: {
                     'Authorization': 'Bearer: ' + user?.token 
                 },
             });
             if (response.ok) {
-                console.log("Event deleted successfully");
+                getEvents();
             } else {
                 console.error(`Failed to delete event: ${response.statusText}`);
             }
@@ -188,30 +210,35 @@ export default function EventList() {
     }
 
     return (
-        <div className="event-list">
-            {events.length === 0 ? 
-            (<p>No events to display...</p>) : events.map(event => (
-                <div key={event._id} className="event-card">
-                    <div className="event-header">
-                        <div className="host-info">
-                            <img src={(event.host.image) ? ("http://localhost:3000/images/profile_pictures/" + event.host.image) : ("/defaultProfilePicture.png")} alt={event.host.username} className="host-image" />
-                            <p className="host-name">{event.host.username}</p>
+        <div className="main-container">
+            <div>
+                <MainMap events={events} deleteEvent={deleteEvent} followEvent={followEvent} following={following} formatDateString={formatDateString} />
+            </div>
+            <div className="event-list">
+                {events.length === 0 ? 
+                (<p>No events to display...</p>) : events.map(event => (
+                    <div key={event._id} className="event-card">
+                        <div className="event-header">
+                            <div className="host-info">
+                                <img src={(event.host.image) ? ("http://localhost:3000/images/profile_pictures/" + event.host.image) : ("/defaultProfilePicture.png")} alt={event.host.username} className="host-image" />
+                                <p className="host-name">{event.host.username}</p>
+                            </div>
+                            <div className="header-right-side">
+                                <p className="followers-display">{(user && following.includes(event._id) && !event.followers.includes(user._id)) ? event.followers.length + 1 : (user && !following.includes(event._id) && event.followers.includes(user._id)) ? event.followers.length - 1 : event.followers.length} <FontAwesomeIcon icon={faUser} /></p>
+                                <p className="event-activity">{event.activity} <FontAwesomeIcon icon={event.activity === "nogomet" ? faSoccerBall : faPersonRunning}/></p>
+                                <button onClick={e => followEvent(event, e)} className="follow-button">{(user && (following.includes(event._id) || event.followers.includes(user._id))) ? "Sledim ✓" : "Sledi +"}</button>
+                                {(user && (user.isAdmin || user._id === event.host._id)) ? (<button onClick={() => deleteEvent(event._id)} className="delete-event"><FontAwesomeIcon icon={faTrash} /></button>) : (<></>)}
+                            </div>
                         </div>
-                        <div className="header-right-side">
-                            <p className="followers-display">{(user && following.includes(event._id) && !event.followers.includes(user._id)) ? event.followers.length + 1 : (user && !following.includes(event._id) && event.followers.includes(user._id)) ? event.followers.length - 1 : event.followers.length} <FontAwesomeIcon icon={faUser} /></p>
-                            <p className="event-activity">{event.activity} <FontAwesomeIcon icon={event.activity === "nogomet" ? faSoccerBall : faPersonRunning}/></p>
-                            <button onClick={e => followEvent(event, e)} className="follow-button">{(user && (following.includes(event._id) || event.followers.includes(user._id))) ? "Sledim ✓" : "Sledi +"}</button>
-                            {(user && (user.isAdmin || user._id === event.host._id)) ? (<button onClick={() => deleteEvent(event)} className="delete-event"><FontAwesomeIcon icon={faTrash} /></button>) : (<></>)}
+                        <div className="event-info">
+                            <h3 className="event-title">{event.name}</h3>
+                            <p className="event-description">{event.description}</p>
+                            <p className="event-date">Dogodek se prične <strong>{formatDateString(event.date)}</strong> ob <strong>{event.time}</strong></p>
                         </div>
+                        <MapComponent title={event.name} location={new LatLng(event.location.coordinates[1], event.location.coordinates[0])}/>
                     </div>
-                    <div className="event-info">
-                        <h3 className="event-title">{event.name}</h3>
-                        <p className="event-description">{event.description}</p>
-                        <p className="event-date">Dogodek se prične <strong>{formatDateString(event.date)}</strong> ob <strong>{event.time}</strong></p>
-                    </div>
-                    <MapComponent title={event.name} location={new LatLng(event.location.coordinates[1], event.location.coordinates[0])}/>
-                </div>
-            ))}
+                ))}
+            </div>
         </div>
     );
